Refactor mostly there
