+++ base64urlsafedata/Cargo.toml
@@ -1,8 +1,10 @@
 [package]
-name = "webauthn-rs"
+name = "base64urlsafedata"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+serde = { version = "1", features = ["derive"] }
+base64 = "0.13"
