+++ webauthn-rs-core/src/macros.rs
