+++ webauthn-rs-demo-wasm/Cargo.toml
@@ -9,8 +9,8 @@ edition = "2018"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-webauthn-rs-demo-shared = { path = "../webauthn-rs-demo-shared" }
-webauthn-rs = { path = "../webauthn-rs", default-features = false, features = ["wasm"] }
+webauthn-rs-demo-shared = { path = "../webauthn-rs-demo-shared", default-features = false }
+webauthn-rs-proto = { path = "../webauthn-rs-proto", default-features = false, features = ["wasm"] }
 wasm-bindgen = { version = "0.2", features = ["serde-serialize"] }
 wasm-bindgen-futures = { version = "0.4" }
 yew-router = "0.16.0"
