+++ webauthn-rs-demo/Cargo.toml
@@ -11,7 +11,8 @@ readme = "README.md"
 license = "MPL-2.0"
 
 [dependencies]
-webauthn-rs-demo-shared = { path = "../webauthn-rs-demo-shared" }
+webauthn-rs-demo-shared = { path = "../webauthn-rs-demo-shared", features = ["core"] }
+webauthn-rs-core = { path = "../webauthn-rs-core" }
 webauthn-rs = { path = "../webauthn-rs" }
 tide = "0.16"
 tide-rustls = "0.3"
